Simplify RequestFactory and end-to-end tests
Introduces a new module: fahrschein-http-test-support
