+++ fahrschein-e2e-test/pom.xml
@@ -35,7 +35,7 @@
         </dependency>
         <dependency>
             <groupId>org.zalando</groupId>
-            <artifactId>fahrschein-http-api</artifactId>
+            <artifactId>fahrschein-http-simple</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
