--- fahrschein-e2e-test/src/test/java/org/zalando/fahrschein/http/simple/SimpleNakadiClientTest.java
@@ -1,12 +1,0 @@
-package org.zalando.fahrschein.http.simple;
-
-import org.zalando.fahrschein.http.AbstractRequestFactoryTest;
-import org.zalando.fahrschein.http.api.ContentEncoding;
-import org.zalando.fahrschein.http.api.RequestFactory;
-
-public class SimpleNakadiClientTest extends AbstractRequestFactoryTest {
-    @Override
-    protected RequestFactory getRequestFactory() {
-        return new SimpleRequestFactory(ContentEncoding.GZIP);
-    }
-}
