--- fahrschein-e2e-test/src/test/java/org/zalando/fahrschein/http/spring/SpringNakadiClientTest.java
@@ -1,30 +1,0 @@
-package org.zalando.fahrschein.http.spring;
-
-import okhttp3.OkHttpClient;
-import okhttp3.logging.HttpLoggingInterceptor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.client.OkHttp3ClientHttpRequestFactory;
-import org.zalando.fahrschein.http.AbstractRequestFactoryTest;
-import org.zalando.fahrschein.http.api.ContentEncoding;
-import org.zalando.fahrschein.http.api.RequestFactory;
-
-public class SpringNakadiClientTest extends AbstractRequestFactoryTest {
-
-    private static final Logger logger = LoggerFactory.getLogger("okhttp3.wire");
-    private static final HttpLoggingInterceptor loggingInterceptor = new HttpLoggingInterceptor(logger::debug);
-    static {
-        loggingInterceptor.setLevel(HttpLoggingInterceptor.Level.HEADERS);
-    }
-
-    @Override
-    protected RequestFactory getRequestFactory() {
-        final OkHttpClient client = new OkHttpClient.Builder()
-                .addInterceptor(loggingInterceptor)
-                .build();
-
-        final OkHttp3ClientHttpRequestFactory clientHttpRequestFactory = new OkHttp3ClientHttpRequestFactory(client);
-        return new SpringRequestFactory(clientHttpRequestFactory, ContentEncoding.GZIP);
-    }
-
-}
