+++ fahrschein-example/pom.xml
@@ -33,7 +33,7 @@
         </dependency>
         <dependency>
             <groupId>org.zalando</groupId>
-            <artifactId>fahrschein-http-api</artifactId>
+            <artifactId>fahrschein-http-simple</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
