+++ fahrschein-http-test-support/pom.xml
@@ -10,9 +10,9 @@
         <version>0.21.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
-    <artifactId>fahrschein-test-support</artifactId>
+    <artifactId>fahrschein-http-test-support</artifactId>
 
-    <name>Fahrschein Test Support</name>
+    <name>Fahrschein HTTP Client Test Support</name>
 
     <build>
         <plugins>
@@ -27,8 +27,9 @@
     <dependencies>
         <dependency>
             <groupId>org.zalando</groupId>
-            <artifactId>fahrschein</artifactId>
+            <artifactId>fahrschein-http-api</artifactId>
             <version>${project.version}</version>
+            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>org.springframework</groupId>
