+++ fahrschein/pom.xml
@@ -17,12 +17,13 @@
     <dependencies>
         <dependency>
             <groupId>org.zalando</groupId>
-            <artifactId>fahrschein-http-api</artifactId>
+            <artifactId>fahrschein-http-simple</artifactId>
             <version>${project.version}</version>
+            <scope>runtime</scope>
         </dependency>
         <dependency>
             <groupId>org.zalando</groupId>
-            <artifactId>fahrschein-http-simple</artifactId>
+            <artifactId>fahrschein-http-api</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
